+++ changelog.rst
@@ -1,11 +1,12 @@
-TBD
-===
+3.4.1 (2022/03/19)
+==================
 
 Bug fixes:
 ----------
 
 * Fix the bug with Redshift not displaying wor count in status ([related issue](https://github.com/dbcli/pgcli/issues/1320)).
 * Show the error status for CSV output format. 
+
 
 3.4.0 (2022/02/21)
 ==================
@@ -50,6 +51,7 @@ Features:
   `destructive_warning` setting to `all|moderate|off`, vs `true|false`. (#1239)
 * Skip initial comment in .pg_session even if it doesn't start with '#'
 * Include functions from schemas in search_path. (`Amjith Ramanujam`_)
+* Easy way to show explain output under F5
 
 Bug fixes:
 ----------
@@ -1127,3 +1129,4 @@ Improvements:
 .. _`thegeorgeous`: https://github.com/thegeorgeous
 .. _`laixintao`: https://github.com/laixintao
 .. _`anthonydb`: https://github.com/anthonydb
+.. _`Daniel Kukula`: https://github.com/dkuku